+++ src/Prdb.Fab.Infrastructure/Persistence/Migrations/20260904072513_PreferredDownloadQuality.cs
@@ -1,0 +1,36 @@
+﻿using Microsoft.EntityFrameworkCore.Migrations;
+
+#nullable disable
+
+namespace Prdb.Fab.Infrastructure.Persistence.Migrations
+{
+    /// <inheritdoc />
+    public partial class PreferredDownloadQuality : Migration
+    {
+        /// <inheritdoc />
+        protected override void Up(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.AddColumn<string>(
+                name: "PreferredDownloadQuality",
+                table: "installation",
+                type: "TEXT",
+                nullable: false,
+                defaultValue: "P2160");
+
+            migrationBuilder.UpdateData(
+                table: "installation",
+                keyColumn: "Id",
+                keyValue: 1,
+                column: "PreferredDownloadQuality",
+                value: "P2160");
+        }
+
+        /// <inheritdoc />
+        protected override void Down(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.DropColumn(
+                name: "PreferredDownloadQuality",
+                table: "installation");
+        }
+    }
+}